refactor: internal => pkg/cli/internal
The only user of the internal package was pkg/cli, so better to
enforce this using the standard go package visibility rules.
