+++ pkg/cli/cli.go
@@ -6,8 +6,8 @@ package cli
 import (
 	_ "embed"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
-	"github.com/rbmk-project/rbmk/internal/rootcmd"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/rootcmd"
 	"github.com/rbmk-project/rbmk/pkg/cli/sh"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 )
