+++ pkg/cli/dig/task.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	"github.com/miekg/dns"
-	"github.com/rbmk-project/rbmk/internal/testable"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/testable"
 	"github.com/rbmk-project/rbmk/pkg/common/closepool"
 	"github.com/rbmk-project/rbmk/pkg/dns/dnscore"
 	"github.com/rbmk-project/rbmk/pkg/x/netcore"
