+++ pkg/cli/internal/qa/censor.go
