+++ pkg/cli/internal/qa/event.go
