+++ pkg/cli/internal/qa/registry.go
