+++ pkg/cli/internal/qa/testdata/pkistore/ZG5zLmdvb2dsZQ==/key.pem
