+++ pkg/cli/internal/qa/testdata/pkistore/d3d3LmV4YW1wbGUuY29t/key.pem
