+++ pkg/cli/internal/rootcmd/README.md
