+++ pkg/cli/intro/intro.go
@@ -8,7 +8,7 @@ import (
 	_ "embed"
 	"fmt"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 )
