+++ pkg/cli/ipuniq/ipuniq.go
@@ -14,7 +14,7 @@ import (
 	"net/netip"
 	"strconv"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 	"github.com/spf13/pflag"
 )
