+++ pkg/cli/mkdir/mkdir.go
@@ -9,7 +9,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 	"github.com/spf13/pflag"
 )
