+++ pkg/cli/nc/task.go
@@ -12,7 +12,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/rbmk-project/rbmk/internal/testable"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/testable"
 	"github.com/rbmk-project/rbmk/pkg/common/closepool"
 	"github.com/rbmk-project/rbmk/pkg/x/netcore"
 )
