+++ pkg/cli/pipe/pipe.go
@@ -6,7 +6,7 @@ package pipe
 import (
 	_ "embed"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 )
