+++ pkg/cli/pipe/write.go
@@ -11,7 +11,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 )
