+++ pkg/cli/sh/builtin.go
@@ -9,8 +9,8 @@ import (
 	"io"
 	"log"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
-	"github.com/rbmk-project/rbmk/internal/rootcmd"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/rootcmd"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 	"github.com/rbmk-project/rbmk/pkg/common/fsx"
 	"mvdan.cc/sh/v3/interp"
