+++ pkg/cli/tar/tar.go
@@ -18,7 +18,7 @@ import (
 	"io/fs"
 	"path/filepath"
 
-	"github.com/rbmk-project/rbmk/internal/markdown"
+	"github.com/rbmk-project/rbmk/pkg/cli/internal/markdown"
 	"github.com/rbmk-project/rbmk/pkg/common/cliutils"
 	"github.com/rbmk-project/rbmk/pkg/common/closepool"
 	"github.com/spf13/pflag"
